+++ packages/moonstone/RangePicker/RangePicker.js
@@ -8,8 +8,7 @@
 import kind from '@enact/core/kind';
 import React from 'react';
 
-import PickerCore from '../Picker/PickerCore';
-import PickerItem from '../Picker/PickerItem';
+import {Picker, PickerItem} from '../internal/Picker';
 import SpottablePicker from '../Picker/SpottablePicker';
 
 const digits = (num) => {
@@ -167,10 +166,17 @@ const RangePickerBase = kind({
 		 * assume auto-sizing. `'small'` is good for numeric pickers, `'medium'` for single or short
 		 * word pickers, `'large'` for maximum-sized pickers.
 		 *
-		 * @type {String}
+		 * You may also supply a number. This number will determine the minumum size of the Picker.
+		 * Setting a number to less than the number of characters in your longest value may produce
+		 * unexpected results.
+		 *
+		 * @type {String|Number}
 		 * @public
 		 */
-		width: React.PropTypes.oneOf([null, 'small', 'medium', 'large']),
+		width: React.PropTypes.oneOfType([
+			React.PropTypes.oneOf([null, 'small', 'medium', 'large']),
+			React.PropTypes.number
+		]),
 
 		/**
 		 * Should the picker stop incrementing when the picker reaches the last element? Set `wrap`
@@ -200,9 +206,9 @@ const RangePickerBase = kind({
 	render: ({label, value, ...rest}) => {
 		delete rest.padded;
 		return (
-			<PickerCore {...rest} index={0} value={value}>
+			<Picker {...rest} index={0} value={value} reverse={false}>
 				<PickerItem key={value} marqueeDisabled>{label}</PickerItem>
-			</PickerCore>
+			</Picker>
 		);
 	}
 });